getShow(id)
Function for getting a show given the ID where the IDs are numbers greater than 0. This is all inside the newly added showsData.js file which is exported and added to the index.js. Addition of validation function in validation.js for verifying if a number is positive for the sake of id argument in getShow(id)