searchShowByTitle and showCard
functions for searching a show by the given querystring and a page number (currently pages have 20 results) and for getting the show's name+image for the cards have been made. Currently searchShowByTitle gives back the show's ID, title, url, premiere data, and image of 'medium' size if it exists otherwise it's null. This was to simplify the response down to relevant data on each entry as oppose to getShow(id) which returns ALL info the API provides on any given show.

May need to edit searchShowByTitle if it's decided that we need more or less information for search results