Getting login to work
Firebase, I've learned, does not play nice with trying to create or log in a user considering I was using frontend functions in the backend for doing such. Major reimagining required. This iteration at least can make a user via frontend but doesn't log anything to our mongodb for this user just yet