logging in and out
set up and confirmed logging in, registering, and logging out of an account is all functional. Confirmed that newly made accounts are added to the firebase db and the mongo db has the user present under the id. Home page will show the email of the user. Use this as another example of how to get user data if they're logged in.

Also deleted idValidation from frontend's accountService to stop errors.

Navbar updated to direct the user to the profile and signout page when clicked (was not navigating previously)